Chore: 메뉴별 리뷰 뷰 테스팅 목적 더미데이터 추가 + 디자인 수정
