--- CheckEat-User/Views/Home/FavoriteStoreModalView.swift
@@ -1,79 +1,0 @@
-//
-//  FavoriteStoreModalView.swift
-//  CheckEat-User
-//
-//  Created by 최준영 on 7/25/25.
-//
-
-import SwiftUI
-
-struct FavoriteStoreModalView: View {
-    @Binding var isPresented: Bool
-    @EnvironmentObject var viewModel: StoreMapViewModel
-    @EnvironmentObject var foodReviewViewModel: FoodReviewViewModel
-    @State private var selectedStore: Store? = nil
-    
-    @State var runTime: String = "09:00 ~ 18:00"
-
-    var body: some View {
-        VStack(alignment: .leading) {
-            HStack {
-                Text("⭐ 즐겨찾기한 가게 (\(viewModel.favoriteStores.count)곳)")
-                    .regular16()
-                Spacer()
-                Button(viewModel.markerMode == .all ? "🔁 즐겨찾기 모드로 전환" : "🔁 전체 보기로 전환") {
-                    viewModel.markerMode = (viewModel.markerMode == .all) ? .favorite : .all
-                    isPresented = false
-                }
-                .regular14()
-            }
-            .padding()
-
-            List(viewModel.favoriteStores, id: \.storeId) { store in
-                Button {
-                    selectedStore = store
-                } label: {
-                    HStack(spacing: 8) {
-                        AsyncImage(url: URL(string: store.sto_img ?? "")) { image in
-                            image
-                                .resizable()
-                                .aspectRatio(1, contentMode: .fit)
-                                .frame(maxWidth: 70)
-                                .cornerRadius(8)
-                        } placeholder: {
-                            ProgressView()
-                                .frame(width: 70, height: 70)
-                        }
-
-                        VStack(alignment: .leading, spacing: 4) {
-                            Text(store.sto_name)
-                                .bold20()
-                            Group {
-                                HStack {
-                                    Image("Location")
-                                    Text(store.sto_address)
-                                }
-                                HStack {
-                                    Image("Time")
-                                    Text("영업시간 \(runTime)")
-                                }
-                            }
-                            .regular14()
-                            .foregroundColor(.secondary)
-                        }
-                    }
-                }
-                .listRowSeparator(.hidden)
-            }
-            .listStyle(.plain)
-        }
-        .fullScreenCover(item: $selectedStore) { store in
-            StoreDetailView(store: store)
-                .environmentObject(viewModel)
-                .environmentObject(foodReviewViewModel)
-                .onDisappear {
-                    viewModel.updateFavoriteStores()
-                }
-        }
-    }
-}
